Moved files to best fit folder structure and removed unused code
