+++ src/modules/app/pages/App.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import { listenForUser } from '../../../redux/actions/auth/authActions';
 import Header from '../components/Header/Header';
-import Admin from './Admin';
+import Admin from '../../admin/pages/Admin';
 
 type Props = {
   listenForUser: () => any;
@@ -14,17 +14,15 @@ export class App extends Component<Props> {
   componentDidMount() {
     this.props.listenForUser();
   }
+
   render() {
     return (
-      <div>
-        <Router>
-          <div>
-            <Header></Header>
-            <h1>Home!</h1>
-            <Route path="/admin/" component={Admin}></Route>
-          </div>
-        </Router>
-      </div>
+      <Router>
+        <div>
+          <Header></Header>
+          <Route path="/admin/" component={Admin}></Route>
+        </div>
+      </Router>
     );
   }
 }
